requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.2.1
django-leaflet==0.27.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dnspython==1.16.0
Flask==1.1.1
geographiclib==1.50
geopy==1.20.0
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.2.0
pprint==0.1
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.44
requests==2.22.0
scipy==1.4.1
six==1.13.0
sqlparse==0.3.1
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==5.0.1
wrapt==1.11.2
